Switch to single-threaded runtime and minimize deps
Copying bytes between sockets isn't CPU-intensive, so the purely
event-driven single-thread approach is fine, and this way I can actually
read the output of strace to verify that it's making the syscalls that I
intended it to.

If you don't want to see all the syscalls from the dynamic linker at
startup, then statically linking with musl is good too.