Use AsyncWriteExt::shutdown to simplify EOF handling
Huh, I didn't realize that was there.

Also upgrading to a new bugfix release of tokio that fixes a soundness
bug in LocalSet. I don't think this program could be affected by the bug
that was fixed since it doesn't use threads, but it uses LocalSet, so
let's just be safe.

Also re-ran `cargo fmt` because apparently I forgot that last time.