Optionally use hashed hostnames on-disk
This reverts commit 00262a03e3813cdf0ac2f9592c6c31bb6a0e4027, "Work
around 108-byte limit in Unix socket addresses", because it didn't work
if the socket was in a directory accessed via a symlink. I could have
fixed that with `fchdir` but I didn't feel it was worth saving, because
it relied heavily on implementation details of Tokio and futures, and
wasn't thread-safe either. Basically just full of foot-guns.

See the new "Alternative design choices" section at the end of README.md
for more details on the many alternatives which don't work.